+++ ncollide_entities/bounding_volume/mod.rs
@@ -7,7 +7,7 @@ pub use bounding_volume::aabb::{AABB, aabb};
 #[doc(inline)]
 pub use bounding_volume::bounding_sphere::{BoundingSphere, bounding_sphere};
 
-pub use bounding_volume::aabb_utils::{implicit_shape_aabb, point_cloud_aabb};
+pub use bounding_volume::aabb_utils::{support_map_aabb, point_cloud_aabb};
 pub use bounding_volume::aabb_ball::ball_aabb;
 pub use bounding_volume::bounding_sphere_utils::{point_cloud_bounding_sphere_with_center, point_cloud_bounding_sphere};
 pub use bounding_volume::bounding_volume_bvt::BoundingVolumeInterferencesCollector;