+++ ncollide_queries/geometry/distance_internal/composite_shape_against_any.rs
@@ -82,7 +82,7 @@ for CompositeShapeAgainstAnyDistCostFn<'a, P, M, G1, G2>
                              *bv.maxs() + self.msum_shift + self.msum_margin);
 
         // Compute the distance to the origin.
-        Some(msum.distance_to_point(&Identity::new(), &na::origin()))
+        Some(msum.distance_to_point(&Identity::new(), &na::origin(), true))
     }
 
     #[inline]