+++ ncollide_queries/point/point_query.rs
@@ -8,7 +8,7 @@ pub trait PointQuery<P: Point, M> {
 
     /// Computes the minimal distance between a point and `self` transformed by `m`.
     #[inline]
-    fn distance_to_point(&self, m: &M, pt: &P) -> <P::Vect as Vector>::Scalar;
+    fn distance_to_point(&self, m: &M, pt: &P, solid: bool) -> <P::Vect as Vector>::Scalar;
 
     /// Tests if the given point is inside of `self` transformed by `m`.
     #[inline]